+++ src/app/[lang]/tempos/page.tsx
@@ -2,9 +2,8 @@
 
 import clsx from 'clsx';
 import { CheckCircleIcon } from 'lucide-react';
-import { Metadata } from 'next';
 import Image from 'next/image';
-import { Locale, useLocale, useTranslations } from 'next-intl';
+import { useLocale, useTranslations } from 'next-intl';
 import { useState } from 'react';
 import ReCAPTCHA from 'react-google-recaptcha';
 import { Controller, useForm } from 'react-hook-form';
@@ -34,84 +33,6 @@ interface QuoteFormInputs {
   message: string;
   consent: boolean;
   reCaptchaToken: string;
-}
-
-export async function generateMetadata({
-  params,
-}: {
-  params: Promise<{ lang: string }>;
-}): Promise<Metadata> {
-  // Use params.slug for route detection
-  const { lang } = (await params) as {
-    lang: Locale;
-  };
-
-  const titles = {
-    en: 'Tempo Purchase, Installation and Rentals - Locaplus',
-    fr: 'Achat, installation et location d’abris Tempo - Locaplus',
-  };
-
-  const descriptions = {
-    en: 'Tempo shelters for purchase, rental, and installation services in Quebec.',
-    fr: "Abris Tempo à vendre, à louer et services d'installation au Québec.",
-  };
-
-  const keywords = {
-    en: 'tempo rentals, tempo purchase, tempo installation',
-    fr: "location d'abris Tempo, achat d'abris Tempo, installation d'abris Tempo",
-  };
-
-  const canonicalUrls = {
-    en: 'https://www.locaplus.net/en/tempos',
-    fr: 'https://www.locaplus.net/fr/abris-tempo',
-  };
-
-  const imageAlts = {
-    en: 'Tent Rentals for Events',
-    fr: 'Location de tentes pour événements',
-  };
-
-  // Default OG image and description
-  const ogImage = 'https://www.locaplus.net/tempo-display.jpg';
-  const description = descriptions[lang] || descriptions['en'];
-
-  return {
-    title: titles[lang] || titles['en'],
-    description,
-    keywords: keywords[lang] || keywords['en'],
-    robots: 'index, follow',
-    openGraph: {
-      title: titles[lang] || titles['en'],
-      description,
-      url: canonicalUrls[lang] || canonicalUrls['en'],
-      siteName: 'Locaplus',
-      images: [
-        {
-          url: ogImage,
-          width: 1200,
-          height: 630,
-          alt: imageAlts[lang],
-        },
-      ],
-      type: 'website',
-    },
-    twitter: {
-      card: 'summary_large_image',
-      title: titles[lang] || titles['en'],
-      description,
-      images: [ogImage],
-    },
-    icons: {
-      icon: '/favicon.ico',
-    },
-    alternates: {
-      canonical: 'https://www.locaplus.net/tempos',
-      languages: {
-        en: 'https://www.locaplus.net/en',
-        fr: 'https://www.locaplus.net/fr',
-      },
-    },
-  };
 }
 
 export default function TemposPage() {
@@ -269,6 +190,7 @@ export default function TemposPage() {
                   <Image
                     src={tempo.img}
                     alt={tempo.name}
+                    sizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
                     fill
                     className="object-cover rounded-xl"
                   />
